Split data: panel info
